Add pointers to linked list in header page
The header page should keep track of the beginning of the
linked list of non-full pages.

The header page also has a pointer to the last page in
the linked list of non-full pages so it doesn't have to
iterate through the linked list to append to the last
page. Note that this is only used when a delete frees
a previously full page because during an insert, it
is necessary to loop through the linked list of non-full
pages to determine if any pages have enough space.