Reuse pages that were full but then deleted from
Append deleted pages to the end of the linked list. This operation
occurs in O(1) time because the header page keeps a page number which
corresponds to the last page of the linked list of non-full pages.

Consider inserting the page at the front of the linked list or back in
order to minimize the travel distance when iterating through the linked
list. Inserting in order might be costly since that forces us to iterate
through the linked list which could be up to O(n) size if tuples were
deleted from all the previously full linked lists.