Implement setNonNullColumnValue() in PageTuple.java
Make space for new column by deleting existing data using
deleteTupleDataRange() and adding space using insertTupleDataRange().
Update the valueOffsets array to match new offsets. Finally, write
new value to the dbPage.

This should cover the cases of updating a NULL column and a
non-NULL column. Currently the VARCHAR case is special, and is the
reason to remove space and then add new space (which accounts for making
the VARCHAR longer or shorter). This can be wrapped with a function to
make handling other variable length cases easier.

Buffer management has not been implemented.