Fixes related to Assignment 1
The "SHOW STORAGE STATS" command broke parser support for "CREATE TABLE"
properties used in the unit tests.  The parser and the stats-command
support have been adjusted to resolve this.

Significant documentation is added to the Pinnable interface, and also
some docs to the TupleFile interface and HeapTupleFile implementation,
to better explain what is going on with pinning of pages and tuples.

Additionally, an important fix was made against the HeapTupleFile
implementation - getNextTuple() will now work on an unpinned tuple.
This greatly simplifies the appropach for making sure everything is
properly unpinned at appropriate times.

Also improved the logging when unpinned DBPages are leaked, so that it's
easier to tell what part of the query the page was leaked from.

The ProjectNode can report if the project operation is "trivial" or not;
this support was optimized so that the test code only has to run once
for the lifetime of the project node, as opposed to every time the node
is checked.

Miscellaneous changes and improvements to other commands, which are
mostly irrelevant, but which help improve the code.