Implement NestedLoopJoinNode
getTuplesToJoin() increments the tuples until it finds a pair of tuples
that can be joined. It iterates through all the rightChild tuples before
iterating through the leftChild tuples, simulating a nested loop.

To handle antijoins, we have a separate loop that iterates through one
left tuple at a time and checks if there is a right tuple that matches.
If there is, it passes to the next left tuple. If there isn't, then it
returns true.

To handle the rest of the joins, it iterates through in the nested loop
order until it finds a set of tuples that satisfy the predicate and then
joins and returns them in getNextTuple().