Fix compilation error
Change checked exception to unchecked exception since failure should
occur if we try to pass a right outer join to NestedLoopJoinNode.
The problem was caused by changing prepare() in the superclass
PlanNode to throw an exception according to the method signature
which would require changing all of its children. The superclass
and its subclasses only need to indicate in the method signature
if a checked exception is thrown.